Moved all utils to common folder
