+++ backend/app/app/api/api_v1/endpoints/login.py
@@ -6,14 +6,14 @@ from urllib.parse import urlencode
 import requests
 from fastapi import APIRouter, Depends, HTTPException, responses, Query
 
-from app.api.utils.db import get_db
-from app.api.utils.link import form_link
-from app.api.utils.security import get_current_user, process_token
 from app.core import config
 from app.core.security import get_code_retrieve_params, get_token_retrieve_params
 from app.db import find_student, create_student
 from app.models.token import TokenRetrieval
 from app.models.user import TokenUser
+from app.utils.db import get_db
+from app.utils.link import form_link
+from app.utils.security import get_current_user, process_token
 
 router = APIRouter()
