+++ backend/app/app/pages/endpoints/main.py
@@ -1,9 +1,8 @@
 import logging
 
-from fastapi import APIRouter, Request, Depends, responses, status
+from fastapi import APIRouter, Depends, responses
 
-from app.api.utils.security import get_current_user_optional
-from app.pages.utils.jinja import templates
+from app.utils.security import get_current_user_optional
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
