+++ backend/app/app/pages/endpoints/profile.py
@@ -2,12 +2,12 @@ import logging
 
 from fastapi import APIRouter, Request, Depends, responses
 
-from app.api.utils.db import get_db
-from app.api.utils.security import get_current_user_optional
 from app.db import get_student_groups, find_student, get_ongoing_semester, get_brief_hours
 from app.models.user import TokenUser
-from app.pages.utils.auth import logout
-from app.pages.utils.jinja import templates
+from app.utils.auth import logout
+from app.utils.db import get_db
+from app.utils.jinja import templates
+from app.utils.security import get_current_user_optional
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
