--- backend/app/app/pages/utils/db.py
@@ -1,5 +1,0 @@
-from starlette.requests import Request
-
-
-def get_db(request: Request):
-    return request.state.db
